% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\docType{data}
\name{indices}
\alias{indices}
\title{RGB and Multi-spectral Indices}
\format{
A data.frame with 21 rows and 3 variables:
\describe{
  \item{index}{index name}
  \item{eq}{equation}
  \item{band}{Bands used}
}
}
\source{
FieldImageR
}
\usage{
indices
}
\description{
Canopy and Green Leaf Index for a potato trial arranged in a p-rep design.
}
\keyword{datasets}
